+++ src/vumi2/config.py
@@ -64,6 +64,7 @@ class BaseConfig:
     sentry_dsn: str | None = None
     http_bind: str | None = None
     log_level: str = "INFO"
+    middlewares: list[dict] = Factory(list)
 
 
 ConfigCallback = Callable[[Attribute, Iterable[str]], Any]